requirements.txt
opencv-python
ultralytics
python-telegram-bot==20.3
python-dotenv